Changed ID types and fixed small bug on report creation

+++ models/report.go
@@ -21,13 +21,13 @@ const (
 
 //Report represents a transaction report in redcoins
 type Report struct {
-	ID              int                `db:"id" json:"id"`
+	ID              int64              `db:"id" json:"id"`
 	Transaction     TransactionType    `db:"transaction" json:"transaction"`
 	BitCoins        float32            `db:"bitcoins" json:"bit_coins"`
 	Value           float32            `db:"value" json:"value"`
 	Currency        utils.CurrencyType `db:"currency" json:"currency"`
 	TransactionDate time.Time          `db:"transaction_date" json:"transaction_date"`
-	UserID          int                `db:"user_id" json:"user_id"`
+	UserID          int64              `db:"user_id" json:"user_id"`
 }
 
 //ReportModel receives the Database table
